refactor(trpc): #2967 — types dérivés et schémas Zod par router
Supprime `RouterInputs` / `RouterOutputs`, dérivés de l'`AppRouter` entier :
chaque module dérive maintenant son type depuis son propre router via
`inferRouterOutputs<typeof xRouter>`, dans son `types.ts`. `RouterInputs`
n'était importé nulle part.

Scinde les deux fichiers de schémas encore partagés par deux routers
voisins, pour qu'un fichier de schémas ne serve plus qu'un seul router :
`releaseLockSchema` passe de `modules/admin/schemas.ts` à
`modules/admin/declarations/schemas.ts`, et les quatre schémas de verrou
de `modules/declaration-remuneration/schemas.ts` au nouveau
`modules/declaration-remuneration/shared/lock/schemas.ts`, sur le modèle
de `shared/draft/schemas.ts` qui sert déjà `declarationDraft`.

Déplacements purs : aucune règle de validation modifiée, les deux
`releaseLockSchema` homonymes restent distincts (UUID côté backoffice,
chaîne libre côté déclarant).
