+++ packages/app/src/modules/admin/__tests__/schemas.test.ts
@@ -1,10 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import {
-	impersonateSearchSchema,
-	releaseLockSchema,
-	startImpersonateSchema,
-} from "../schemas";
+import { impersonateSearchSchema, startImpersonateSchema } from "../schemas";
 import { updateLockTimeoutSchema } from "../settings/schemas";
 
 describe("admin schemas", () => {
@@ -69,22 +65,5 @@ describe("updateLockTimeoutSchema", () => {
 
 	it("rejects a missing timeout", () => {
 		expect(updateLockTimeoutSchema.safeParse({}).success).toBe(false);
-	});
-});
-
-describe("releaseLockSchema", () => {
-	it("accepts a valid UUID declarationId", () => {
-		const result = releaseLockSchema.safeParse({
-			declarationId: "6ba7b810-9dad-11d1-80b4-00c04fd430c8",
-		});
-		expect(result.success).toBe(true);
-	});
-
-	it.each([
-		"",
-		"not-a-uuid",
-		"123456789",
-	])("rejects a non-UUID declarationId %s", (declarationId) => {
-		expect(releaseLockSchema.safeParse({ declarationId }).success).toBe(false);
 	});
 });
