+++ packages/app/src/modules/admin/index.ts
@@ -10,7 +10,5 @@ export { AdminReferentsPage } from "./referents";
 export {
 	type ImpersonateSearchInput,
 	impersonateSearchSchema,
-	type ReleaseLockInput,
-	releaseLockSchema,
 	sirenSchema,
 } from "./schemas";
