+++ packages/app/src/modules/declaration-remuneration/index.ts
@@ -54,8 +54,9 @@ export {
 export type {
 	DeclarationLockState,
 	LockHolder,
+	LockHolderDisplay,
 	ReadOnlyReason,
-} from "./shared/lock/useDeclarationLock";
+} from "./shared/lock/types";
 export { useDeclarationLock } from "./shared/lock/useDeclarationLock";
 export { ComplianceConfirmation } from "./steps/ComplianceConfirmation";
 export { CompliancePathChoice } from "./steps/CompliancePathChoice";
