+++ packages/app/src/modules/declaration-remuneration/schemas.ts
@@ -196,19 +196,3 @@ export type CategoryFormValues = z.infer<typeof categoryFormSchema>;
 export const saveCompliancePathSchema = z.object({
 	path: z.enum(COMPLIANCE_PATHS),
 });
-
-export const acquireLockSchema = z.object({
-	declarationId: z.string(),
-});
-
-export const heartbeatSchema = z.object({
-	declarationId: z.string(),
-});
-
-export const releaseLockSchema = z.object({
-	declarationId: z.string(),
-});
-
-export const getLockStateSchema = z.object({
-	declarationId: z.string(),
-});
