+++ packages/app/src/modules/declaration-remuneration/shared/lock/DeclarationLockAlert.tsx
@@ -1,7 +1,7 @@
-import type { LockHolder } from "./useDeclarationLock";
+import type { LockHolderDisplay } from "./types";
 
 type DeclarationLockAlertProps = {
-	holder: Pick<LockHolder, "firstName" | "lastName" | "email">;
+	holder: LockHolderDisplay;
 };
 
 export function DeclarationLockAlert({ holder }: DeclarationLockAlertProps) {
