+++ packages/app/src/modules/declaration-remuneration/shared/lock/schemas.ts
@@ -1,0 +1,17 @@
+import { z } from "zod";
+
+export const acquireLockSchema = z.object({
+	declarationId: z.string(),
+});
+
+export const heartbeatSchema = z.object({
+	declarationId: z.string(),
+});
+
+export const releaseLockSchema = z.object({
+	declarationId: z.string(),
+});
+
+export const getLockStateSchema = z.object({
+	declarationId: z.string(),
+});
