+++ packages/app/src/modules/declaration-remuneration/shared/lock/useDeclarationLock.ts
@@ -8,20 +8,8 @@ import {
 	DECLARATION_LOCK_CONFLICT_MESSAGE,
 	LOCK_HEARTBEAT_INTERVAL_MS,
 } from "~/modules/domain";
-import { api, type RouterOutputs } from "~/trpc/react";
-
-export type LockHolder = NonNullable<
-	RouterOutputs["declarationLock"]["getLockState"]["holder"]
->;
-
-export type ReadOnlyReason = "impersonation" | "modification_closed" | "lock";
-
-export type DeclarationLockState = {
-	isReadOnly: boolean;
-	reason: ReadOnlyReason | null;
-	holder: LockHolder | null;
-	isLoading: boolean;
-};
+import { api } from "~/trpc/react";
+import type { DeclarationLockState, LockHolder, ReadOnlyReason } from "./types";
 
 type UseDeclarationLockOptions = {
 	declarationId: string;
